app_to_postgres/requirements.txt
flask
paramiko
sshtunnel
psycopg2-binary
matplotlib
pandas
